fix: address CodeRabbit review comments
- usePendingCalls: track delayed setTimeout and clear it in teardown
  so back-to-back start/stop runs don't leak the lingering tick
- Move ToolCallHistoryItem to src/types/toolCallHistory.ts so domain
  types are no longer owned by a Vue component
- test_scrollable: save/restore globalThis.getComputedStyle in
  afterEach to prevent test pollution across files
